add test function for array and slices complete
